Refactor: Pull Miles Into Object
Pull miles code into its own object.

The base code was used in a demo before students had learned about variables and functions.  We don't need to worry about that here, so can add variables for location, etc.

The delta variables aren't used yet, but will be in the next commit when adding the animations.