Start Scaling Button: Add Functionality
Add bounding box logic to check if we are clicking in the 'Start Scaling' button.

If we are toggle isScaling to change the button text and pause the scaling animation.

I also added a globalScale variable because I'll probably want to be able to change the canvas size while demoing to make it easier to view.  It isn't being applied yet, but I wanted to start referencing the globalScale variable now while adding the bounding box logic so I won't need to edit this logic later.