Add Rotate Button
Add rotate button (UI and functionality).

I wanted the button to mirror the scaling one, so added in a constant for the canvas width / height (planning on keeping the canvas square to avoid distortion).  We can use this constant to anchor the button to the right side.  I don't want to use the built in width and height variables since referencing it before setup runs.